Add support for rolling `Deployments` owned by unknown CRs, like the case of Crossplane providers
In that case, the Crossplane provider `Deployments` are owned by a `ProviderRevision`, which in turn is owned by a `Provider`. The restarted CronJob was failing on these cases.

Now the restarter will find the "last known" owner in the ownership chain, instead of the top-most one, and try to roll that one.